pm-agent/pyproject.toml
[project]
name = "pm-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.75.0",
    "python-dotenv>=1.2.1",
]
